Added stability, and added sponsors to inference interface (#2383)
